requirements.txt
torch>=1.4.0
pandas>=1.0.3
h5py>=2.10.0
pycocotools>=2.0.0
scikit-learn>=0.22.1
torchvision>=0.4.0
numpy>=1.17.3
Pillow>=7.0.0
matplotlib>=3.2.1
scikit-image>=0.16.2
configparser>=5.0.0
tqdm>=4.41.1
imgaug>=0.4.0
opencv-python>=4.2.0.32
mlflow>=1.7.2
#Hydra>=2.5
torchsummary>=1.5.1
minio>=5.0.8
scipy~=1.5.0
setuptools~=46.1.3
bokeh~=2.1.1
torch~=1.6.0
Hydra~=2.5
PyYAML~=5.3.1
sklearn~=0.0
kornia>=0.4.0
Shapely~=1.7.1
imageio~=2.9.0
line_profiler~=3.1.0